[App] feat: restructure code base
